requirements.txt
cmake
torch
torchvision
numpy
pillow
maraboupy==2.0.0
